fix: Client and server missues on utils session function. Guest card png.
