+++ apps/main-landing/src/app/creators/utils/index.ts
@@ -1,7 +1,5 @@
 import { Hex } from 'viem';
 import { CHAIN, CREATOR_API_URL } from '@idriss-xyz/constants';
-import { getAccessToken, User } from '@privy-io/react-auth';
-import { SetStateAction } from 'react';
 
 type BrowserBasedImageProperties = {
   svgSrc: string;
@@ -239,22 +237,4 @@ export const getChainIdsFromShortNames = (shortNames: string[]) => {
 };
 
 export { useStartEarningNavigation } from './navigation';
-
-export const setCreatorIfSessionPresent = async (
-  user: User,
-  setCreator: (value: SetStateAction<CreatorProfileResponse | null>) => void,
-) => {
-  const walletAddress = user.wallet?.address as Hex | undefined;
-  const authToken = await getAccessToken();
-  if (!authToken || !user.id) {
-    throw new Error('Could not get auth token or user ID for new user.');
-  }
-  const fetchedCreator = await getCreatorProfile(authToken);
-  if (fetchedCreator == null || fetchedCreator === undefined) {
-    console.error(
-      `Could not find creator with address ${walletAddress}, this may be a new user.`,
-    );
-    return;
-  }
-  setCreator(fetchedCreator);
-};
+export { setCreatorIfSessionPresent } from './session';
